+++ lib/regexp_parser/expression/classes/conditional.rb
@@ -31,7 +31,7 @@ module Regexp::Expression
         expressions.last << exp
       end
 
-      def add_sequence(active_opts = {})
+      def add_sequence(active_opts = Regexp::Options.new)
         raise TooManyBranches.new if branches.length == 2
         params = { conditional_level: conditional_level + 1 }
         Branch.add_to(self, params, active_opts)