+++ lib/regexp_parser/expression/sequence_operation.rb
@@ -14,7 +14,7 @@ module Regexp::Expression
       expressions.last << exp
     end
 
-    def add_sequence(active_opts = {})
+    def add_sequence(active_opts = Regexp::Options.new)
       self.class::OPERAND.add_to(self, {}, active_opts)
     end